requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
email-validator
python-jose[cryptography]
cryptography
httpx
certifi
beautifulsoup4
feedparser
pillow
python-multipart
google-genai
openai
authlib
itsdangerous
pg8000
bcrypt
